requirements.txt

torch
flask
requests
sentence-transformers
huggingface_hub
numpy
faiss-cpu
torch
scikit-learn
flask-cors
transformers
langchain

